* enumerator.c (inspect_enumerator): should use long.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@23497 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ enumerator.c
@@ -605,7 +605,7 @@ inspect_enumerator(VALUE obj, VALUE dummy, int recur)
     rb_str_buf_cat2(str, rb_id2name(e->meth));
 
     if (e->args) {
-	int    argc = RARRAY_LEN(e->args);
+	long   argc = RARRAY_LEN(e->args);
 	VALUE *argv = RARRAY_PTR(e->args);
 
 	rb_str_buf_cat2(str, "(");
